--Unit of Work Pattern Added. --Database and relationships between tables established. --Microsoft.AspNetCore.Identity utilized.
--Seed Data
Currently, the model builder can not use the seed data for the test users. This is because the foreign key of the collections table requires the id value of the Users table; the id value in the user's table will be created once the UserManager is called to create the user and insert it into the table, meaning unless the user exists data for the other tables can not be inserted. To address this, I plan to implement a way to seed the data within the CreateAdminUserAndTestUsers() function of the Collectors Hub Context. This method ensures the data is seeded only once as long as the test user exists inside the Users table.

--Model's
Models are a work in progress and still need data validation

--View Models
View Models are also a work in progress

--View's
all of the Views other than the item views are functional but do not incorporate a layout or styling.

--Nav bar
still needs to be built

--website functionality
users can register accounts, log in, and log out. Users not logged in can view other users' collections and filter results by a partial or full user name and the tag associated with that collection. Logged users have the same privileges but can view their collections, add a collection, or edit them.