--Modified Collecion Controller to use temp data for it's Add/Edit IAction --Item Controller, Add/Edit IAction, and view
For this section, I added the Add/Edit option and view. The view takes the item inside the model and calls the image converter to convert the byte array into a base 64 image string in the Add/Edit page to display to the user. At first, there isn't an image in the add section until the user attempts to upload an image; if the add operation or edit operations fail due to validation, the image data remains as long as the user doesn't overwrite it with a new image.