--Added Admin Area The admin can delete users, collections, and items
--Modified Nav Bar
I Modified the Nav Bar of the layout.cshtml so that if the admin is logged in, the admin Tools Link is displayed.

--Fixed Error in Seed Data for fake users
Initially, I set the collection ID for the seed data to a default value to get the collection that was inserted into the database. However, due to the autoincrement, this only works the first time the database is created, which makes the application challenging to test. It now gets the collection that matches the inserted data.

--Added filtering for the Collection View Model for the item names

--removed users from the UserViewModels GetUsers Method by removing users that do not have a collection

--CSS.